Allow https CSP headers
*Problem*
If foreman_cockpit is installed in a system with https, CSP will only
enable the http version of it to be loaded, and since we require https
to only load https content in its page, it'll be blocked.

*Solution*
Allow https headers if we detect foreman is running in https mode
(require_ssl?).